+++ index.d.ts
@@ -1,0 +1,8 @@
+export {
+  default as CheckboxFilterAccordion,
+  CheckboxFilterAccordionProps,
+} from './lib/CheckboxFilterAccordion';
+export {
+  default as DynamicSelectionFilterAccordion,
+  DynamicSelectionFilterAccordionProps,
+} from './lib/DynamicSelectionFilterAccordion';