+++ lib/DynamicSelectionFilterAccordion/index.d.ts
@@ -1,0 +1,4 @@
+export {
+  default,
+  DynamicSelectionFilterAccordionProps,
+} from './DynamicSelectionFilterAccordion';